move Counter.vue and axios_sample to components directory
