+++ src/main.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import App from './App.vue'
-import Counter from './Counter.vue'
-import AxiosSample from './axios_sample/AxiosSample.vue'
+import Counter from './components/Counter.vue'
+import AxiosSample from './components/axios_sample/AxiosSample.vue'
 import Forms from './components/Forms.vue'
 
 Vue.config.productionTip = false
